% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_version.R
\name{glm_version}
\alias{glm_version}
\title{Return the current GLM model version}
\usage{
glm_version(as_char = FALSE)
}
\arguments{
\item{as_char}{return glm version as a character string? (TRUE/FALSE)}
}
\description{
Returns the current version of the GLM model being used
}
\examples{
print(glm_version())


}
\author{
Luke Winslow, Jordan Read
}
\keyword{methods}
