% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_glm.R
\name{run_glm}
\alias{run_glm}
\title{run the GLM model}
\usage{
run_glm(sim_folder = ".", verbose = TRUE, system.args = character())
}
\arguments{
\item{sim_folder}{the directory where simulation files are contained}

\item{verbose}{Logical: Should output of GLM be shown}

\item{system.args}{Optional arguments to pass to GLM executable}
}
\description{
This runs the GLM model on the specific simulation stored in \code{sim_folder}. 
The specified \code{sim_folder} must contain a valid NML file.
}
\examples{
sim_folder <- system.file('extdata', package = 'GLMr')
run_glm(sim_folder)
\dontrun{
out_file <- file.path(sim_folder,'Output/output.nc')
nml_file <- file.path(sim_folder,'glm3.nml')
library(glmtools)
plot_temp(nc_file = out_file)
cat('find plot here: '); cat(fig_path)
}
}
\author{
Jordan Read, Luke Winslow, Hilary Dugan
}
\keyword{methods}
